feat: More utils
