+++ src/ui/base.ts
@@ -6,6 +6,7 @@
 //
 
 import { create } from "./core";
+import { mount, render } from "./dom";
 
 type HTMLComponents = {
   [K in keyof HTMLElementTagNameMap]: (
@@ -34,3 +35,23 @@ export const html = new Proxy({} as HTMLComponents, {
     return (props: Parameters<typeof create>[1]) => create(type, props);
   },
 });
+
+/**
+ * Build a component from a function.
+ * @param fn The function to build the component.
+ * @returns The component.
+ */
+export function component<
+  T extends { [k: string | symbol | number]: unknown } = {},
+  K extends HTMLElement = HTMLElement
+>(fn: (props: T) => K) {
+  return {
+    view: fn,
+    render: (props: T, parent: HTMLElement) => {
+      render(fn(props), parent);
+    },
+    mount: (props: T, parent: HTMLElement) => {
+      mount(fn(props), parent);
+    },
+  };
+}
